Alpha IO/requirements.txt
openai
faiss-cpu
tweepy
yfinance
pandas
numpy
scikit-learn
streamlit
flask
web3
python-dotenv
plotly
